+++ angular-upload.js
@@ -14,6 +14,7 @@ angular.module('lr.upload.directives').directive('uploadButton', [
       scope: {
         options: '=?uploadButton',
         multiple: '=?',
+        accept: '=?',
         forceIFrameUpload: '=?forceIframeUpload',
         url: '@',
         method: '@',
@@ -33,6 +34,9 @@ angular.module('lr.upload.directives').directive('uploadButton', [
         });
         var fileInput = angular.element('<input type="file" />');
         fileInput.on('change', function uploadButtonFileInputChange() {
+          if (fileInput[0].files && fileInput[0].files.length === 0) {
+            return;
+          }
           if (!scope.options) {
             scope.options = {};
           }
@@ -51,9 +55,12 @@ angular.module('lr.upload.directives').directive('uploadButton', [
             scope.onComplete({ response: response });
           });
         });
+        scope.$watch('accept', function uploadButtonAcceptWatch(value) {
+          fileInput.attr('accept', angular.isArray(value) ? value.join(',') : value);
+        });
         el.append(fileInput);
         if (upload.support.formData) {
-          scope.$watch('multiple + forceIFrameUpload', function uploadButtonWatch(value) {
+          scope.$watch('multiple + forceIFrameUpload', function uploadButtonMultipleWatch(value) {
             fileInput.attr('multiple', !!(value && !scope.forceIFrameUpload));
           });
         }
@@ -73,12 +80,14 @@ angular.module('lr.upload.formdata', []).factory('formDataTransform', function (
             files.push(file);
           });
         });
-        if (files.length > 1) {
-          angular.forEach(files, function (file, index) {
-            formData.append(key + '[' + index + ']', file);
-          });
-        } else {
-          formData.append(key, files[0]);
+        if (files.length !== 0) {
+          if (files.length > 1) {
+            angular.forEach(files, function (file, index) {
+              formData.append(key + '[' + index + ']', file);
+            });
+          } else {
+            formData.append(key, files[0]);
+          }
         }
       } else {
         formData.append(key, value);