+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 1.0.5
+ - uploadButton
+  - Only upload if any files where selected
+  - accept attribute (https://developer.mozilla.org/en-US/docs/Web/HTML/Element/Input#attr-accept)
+  	Also allows for an js array as input
+
+# 1.0.4
+ - empty input fix
+
+# 1.0.3
+ - bump test
+
 # 1.0.2
  - Don't replace the clone when done uploading, since you usually want the input to be cleared.
  - Added onComplete that fires for both success and failure